Adapt to IB API version 9.72+

+++ core/src/main/java/com/highpowerbear/hpbanalytics/dao/IbLoggerDao.java
@@ -19,5 +19,5 @@ public interface IbLoggerDao {
     List<IbOrder> getOpenIbOrders(IbAccount ibAccount);
     void newIbOrder(IbOrder ibOrder);
     void updateIbOrder(IbOrder ibOrder);
-    IbOrder getIbOrderByPermId(IbAccount ibAccount, Integer permId);
+    IbOrder getIbOrderByPermId(IbAccount ibAccount, Long permId);
 }
